--- src/Sylius/Bundle/CartBundle/Form/Type/CartType.php
@@ -1,45 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Sylius\Bundle\CartBundle\Form\Type;
-
-use Sylius\Bundle\ResourceBundle\Form\Type\AbstractResourceType;
-use Symfony\Component\Form\FormBuilderInterface;
-
-/**
- * Cart form form.
- * It is built from collection of cart items forms.
- *
- * @author Paweł Jędrzejewski <pawel@sylius.org>
- */
-class CartType extends AbstractResourceType
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder
-            ->add('items', 'collection', [
-                'type' => 'sylius_cart_item',
-            ])
-            ->add('notes')
-        ;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return 'sylius_cart';
-    }
-}
