--- src/Sylius/Bundle/CartBundle/Resources/config/doctrine/model/CartItem.orm.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Paweł Jędrzejewski
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
-                                      http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <mapped-superclass name="Sylius\Component\Order\Model\OrderItem" table="sylius_cart_item">
-    </mapped-superclass>
-
-</doctrine-mapping>
