--- src/Sylius/Bundle/CartBundle/Resources/config/services.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Paweł Jędrzejewski
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services
-                               http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <parameters>
-        <parameter key="sylius.listener.session_cart.class">Sylius\Bundle\CartBundle\EventListener\SessionCartSubscriber</parameter>
-
-        <parameter key="sylius.cart.purger.class">Sylius\Bundle\CartBundle\Purger\ExpiredCartsPurger</parameter>
-    </parameters>
-
-    <services>
-        <service id="sylius.cart.cart_modifier" class="Sylius\Component\Order\Modifier\CartModifier">
-            <argument type="service" id="sylius.order_processing.order_processor" />
-            <argument type="service" id="sylius.order_item_quantity_modifier" />
-        </service>
-        <service id="sylius.context.cart_new" class="Sylius\Component\Order\Context\CartContext">
-            <argument type="service" id="sylius.factory.cart" />
-            <tag name="sylius.cart_context" priority="-999" />
-        </service>
-        <service id="sylius.context.cart_session_based" class="Sylius\Bundle\CartBundle\Context\SessionBasedCartContext">
-            <argument type="service" id="session" />
-            <argument>_sylius.cart</argument>
-            <argument type="service" id="sylius.repository.cart" />
-            <tag name="sylius.cart_context" priority="-777" />
-        </service>
-        <service id="sylius.context.cart" class="Sylius\Component\Order\Context\CompositeCartContext" />
-
-        <service id="sylius.listener.session_cart" class="%sylius.listener.session_cart.class%">
-            <argument type="service" id="sylius.context.cart" />
-            <argument>_sylius.cart</argument>
-            <tag name="kernel.event_subscriber" />
-        </service>
-
-        <service id="sylius.cart.purger" class="%sylius.cart.purger.class%">
-            <argument type="service" id="sylius.manager.cart" />
-            <argument type="service" id="sylius.repository.cart" />
-        </service>
-    </services>
-</container>
