--- src/Sylius/Bundle/CartBundle/Resources/config/templating.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Paweł Jędrzejewski
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services
-                               http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <parameters>
-        <parameter key="sylius.templating.helper.cart.class">Sylius\Bundle\CartBundle\Templating\Helper\CartHelper</parameter>
-    </parameters>
-
-    <services>
-        <service id="sylius.templating.helper.cart" class="%sylius.templating.helper.cart.class%" lazy="true">
-            <argument type="service" id="sylius.context.cart" />
-            <argument type="service" id="sylius.factory.cart_item" />
-            <argument type="service" id="form.factory" />
-            <argument type="service" id="sylius.order_item_quantity_modifier" />
-            <tag name="templating.helper" alias="sylius_cart" />
-        </service>
-    </services>
-
-</container>
