--- src/Sylius/Bundle/CartBundle/Resources/config/twig.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Paweł Jędrzejewski
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services
-                               http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <parameters>
-        <parameter key="sylius.twig.extension.cart.class">Sylius\Bundle\CartBundle\Twig\CartExtension</parameter>
-    </parameters>
-
-    <services>
-        <service id="sylius.twig.extension.cart" class="%sylius.twig.extension.cart.class%" public="false">
-            <argument type="service" id="sylius.templating.helper.cart" />
-            <tag name="twig.extension" />
-        </service>
-    </services>
-
-</container>
