--- src/Sylius/Bundle/CartBundle/spec/Form/Type/CartTypeSpec.php
@@ -1,66 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace spec\Sylius\Bundle\CartBundle\Form\Type;
-
-use PhpSpec\ObjectBehavior;
-use Symfony\Component\Form\FormBuilder;
-use Symfony\Component\Form\FormTypeInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-/**
- * @author Paweł Jędrzejewski <pawel@sylius.org>
- */
-final class CartTypeSpec extends ObjectBehavior
-{
-    function let()
-    {
-        $this->beConstructedWith('Cart', ['sylius']);
-    }
-
-    function it_is_initializable()
-    {
-        $this->shouldHaveType('Sylius\Bundle\CartBundle\Form\Type\CartType');
-    }
-
-    function it_is_a_form_type()
-    {
-        $this->shouldImplement(FormTypeInterface::class);
-    }
-
-    function it_builds_form_with_items_collection(FormBuilder $builder)
-    {
-        $builder
-            ->add('items', 'collection', ['type' => 'sylius_cart_item'])
-            ->willReturn($builder)
-        ;
-
-        $builder
-            ->add('notes')
-            ->willReturn($builder)
-        ;
-
-        $this->buildForm($builder, []);
-    }
-
-    function it_defines_assigned_data_class(OptionsResolver $resolver)
-    {
-        $resolver
-            ->setDefaults([
-                'data_class' => 'Cart',
-                'validation_groups' => ['sylius'],
-            ])
-            ->shouldBeCalled()
-        ;
-
-        $this->configureOptions($resolver);
-    }
-}
