+++ src/Sylius/Bundle/OrderBundle/Context/SessionBasedCartContext.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\CartBundle\Context;
+namespace Sylius\Bundle\OrderBundle\Context;
 
 use Sylius\Component\Order\Context\CartContextInterface;
 use Sylius\Component\Order\Context\CartNotFoundException;
