+++ src/Sylius/Bundle/OrderBundle/Doctrine/ORM/OrderRepository.php
@@ -12,6 +12,7 @@
 namespace Sylius\Bundle\OrderBundle\Doctrine\ORM;
 
 use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
+use Sylius\Component\Order\Model\OrderInterface;
 use Sylius\Component\Order\Repository\OrderRepositoryInterface;
 
 /**
@@ -95,6 +96,20 @@ class OrderRepository extends EntityRepository implements OrderRepositoryInterfa
         ;
 
         return $queryBuilder
+            ->getQuery()
+            ->getOneOrNullResult()
+        ;
+    }
+    /**
+     * {@inheritdoc}
+     */
+    public function findCartById($id)
+    {
+        return $this->createQueryBuilder('o')
+            ->where('o.id = :id')
+            ->andWhere('o.state = :state')
+            ->setParameter('state', OrderInterface::STATE_CART)
+            ->setParameter('id', $id)
             ->getQuery()
             ->getOneOrNullResult()
         ;
