+++ src/Sylius/Bundle/OrderBundle/EventListener/SessionCartSubscriber.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\CartBundle\EventListener;
+namespace Sylius\Bundle\OrderBundle\EventListener;
 
 use Sylius\Component\Order\Context\CartContextInterface;
 use Sylius\Component\Order\Context\CartNotFoundException;
