+++ src/Sylius/Bundle/OrderBundle/Resources/config/templating.xml
@@ -18,12 +18,20 @@
 
     <parameters>
         <parameter key="sylius.templating.helper.adjustment.class">Sylius\Bundle\OrderBundle\Templating\Helper\AdjustmentsHelper</parameter>
+        <parameter key="sylius.templating.helper.cart.class">Sylius\Bundle\OrderBundle\Templating\Helper\CartHelper</parameter>
     </parameters>
 
     <services>
         <service id="sylius.templating.helper.adjustment" class="%sylius.templating.helper.adjustment.class%">
             <argument type="service" id="sylius.adjustments_aggregator" />
             <tag name="templating.helper" alias="sylius_adjustments" />
+        </service>
+        <service id="sylius.templating.helper.cart" class="%sylius.templating.helper.cart.class%" lazy="true">
+            <argument type="service" id="sylius.context.cart" />
+            <argument type="service" id="sylius.factory.cart_item" />
+            <argument type="service" id="form.factory" />
+            <argument type="service" id="sylius.order_item_quantity_modifier" />
+            <tag name="templating.helper" alias="sylius_cart" />
         </service>
     </services>
