+++ src/Sylius/Bundle/OrderBundle/Resources/config/twig.xml
@@ -18,11 +18,16 @@
 
     <parameters>
         <parameter key="sylius.twig.extension.aggregate_adjustments.class">Sylius\Bundle\OrderBundle\Twig\AggregateAdjustmentsExtension</parameter>
+        <parameter key="sylius.twig.extension.cart.class">Sylius\Bundle\OrderBundle\Twig\CartExtension</parameter>
     </parameters>
 
     <services>
         <service id="sylius.twig.extension.aggregate_adjustments" class="%sylius.twig.extension.aggregate_adjustments.class%" public="false">
             <argument type="service" id="sylius.templating.helper.adjustment" />
+            <tag name="twig.extension" />
+        </service>
+        <service id="sylius.twig.extension.cart" class="%sylius.twig.extension.cart.class%" public="false">
+            <argument type="service" id="sylius.templating.helper.cart" />
             <tag name="twig.extension" />
         </service>
     </services>
